requirements.txt
torchvision 
torchaudio 
cudatoolkit==11.3
tensorboard_logger
pandas
transformers